+++ src/main/java/ru/ssau/tk/vaa/LR_Voevodin_Kashapov/ui/SimpleOperationsWindow.java
@@ -261,21 +261,36 @@ public class SimpleOperationsWindow extends JDialog {
         saveButton1.addActionListener(evt -> {
             int returnVal = saveChooser.showSaveDialog(this);
             if (returnVal == JFileChooser.APPROVE_OPTION) {
-                File file = new File(saveChooser.getSelectedFile() + ".bin");
-                try (BufferedOutputStream out = new BufferedOutputStream(new FileOutputStream(file))) {
-                    if (function1 != null) {
-                        FunctionsIO.serialize(out, function1);
-                        JOptionPane.showMessageDialog(this,
-                                "Файл '" + saveChooser.getSelectedFile() +
-                                        ".bin' сохранен");
-                    } else {
-                        throw new IOException();
+                String fileName = saveChooser.getSelectedFile() + ".bin";
+                int flag = 1;
+                File file = new File(fileName);
+                if (file.exists()) {
+                    flag = -1;
+                    int ind = JOptionPane.showConfirmDialog(this, "Файл с таким названием уже существует в данном расположении. Вы хотите сохранить файл с названием " +
+                                    HelperMethods.getFinalNewDestinationFile(new File("output"), file).getName() + "?",
+                            "Предупреждение", JOptionPane.YES_NO_OPTION);
+                    if (ind == 0) {
+                        file = HelperMethods.getFinalNewDestinationFile(new File("output"), file);
+                        flag = 1;
                     }
-                } catch (IOException e) {
-                    new ErrorWindow(this, e);
+                }
+                if (flag != -1){
+                    try (BufferedOutputStream out = new BufferedOutputStream(new FileOutputStream(file))) {
+                        if (function1 != null) {
+                            FunctionsIO.serialize(out, function1);
+                            JOptionPane.showMessageDialog(this,
+                                    "Файл '" + file.getName() +
+                                            " сохранен");
+                        } else {
+                            throw new IOException();
+                        }
+                    } catch (Exception e) {
+                        new ErrorWindow(this, e);
+                    }
                 }
 
             }
+
         });
         downloadButton1.addActionListener(evt -> {
             int returnVal = downloadChooser.showOpenDialog(this);
@@ -308,29 +323,41 @@ public class SimpleOperationsWindow extends JDialog {
             }
         });
 
-        saveButton2.addActionListener(evt ->
-
-        {
+        saveButton2.addActionListener(evt -> {
             int returnVal = saveChooser.showSaveDialog(this);
             if (returnVal == JFileChooser.APPROVE_OPTION) {
-                File file = new File(saveChooser.getSelectedFile() + ".bin");
-                try (BufferedOutputStream out = new BufferedOutputStream(new FileOutputStream(file))) {
-                    if (function2 != null) {
-                        FunctionsIO.serialize(out, function2);
-                        JOptionPane.showMessageDialog(this,
-                                "Файл '" + saveChooser.getSelectedFile() +
-                                        ".bin' сохранен");
-                    } else {
-                        throw new IOException();
+                String fileName = saveChooser.getSelectedFile() + ".bin";
+                int flag = 1;
+                File file = new File(fileName);
+                if (file.exists()) {
+                    flag = -1;
+                    int ind = JOptionPane.showConfirmDialog(this, "Файл с таким названием уже существует в данном расположении. Вы хотите сохранить файл с названием " +
+                                    HelperMethods.getFinalNewDestinationFile(new File("output"), file).getName() + "?",
+                            "Предупреждение", JOptionPane.YES_NO_OPTION);
+                    if (ind == 0) {
+                        file = HelperMethods.getFinalNewDestinationFile(new File("output"), file);
+                        flag = 1;
                     }
-                } catch (IOException e) {
-                    new ErrorWindow(this, e);
                 }
+                if (flag != -1){
+                    try (BufferedOutputStream out = new BufferedOutputStream(new FileOutputStream(file))) {
+                        if (function2 != null) {
+                            FunctionsIO.serialize(out, function2);
+                            JOptionPane.showMessageDialog(this,
+                                    "Файл '" + file.getName() +
+                                            " сохранен");
+                        } else {
+                            throw new IOException();
+                        }
+                    } catch (Exception e) {
+                        new ErrorWindow(this, e);
+                    }
+                }
+
             }
         });
 
         downloadButton2.addActionListener(evt ->
-
         {
             int returnVal = downloadChooser.showOpenDialog(this);
             if (returnVal == JFileChooser.APPROVE_OPTION) {
@@ -376,25 +403,39 @@ public class SimpleOperationsWindow extends JDialog {
             fillMap();
         });
 
-        saveButton3.addActionListener(evt ->
-
-        {
+        saveButton3.addActionListener(evt -> {
             int returnVal = saveChooser.showSaveDialog(this);
             if (returnVal == JFileChooser.APPROVE_OPTION) {
-                File file = new File(saveChooser.getSelectedFile() + ".bin");
-                try (BufferedOutputStream out = new BufferedOutputStream(new FileOutputStream(file))) {
-                    if (function3 != null) {
-                        FunctionsIO.serialize(out, function3);
-                        JOptionPane.showMessageDialog(this,
-                                "Файл '" + saveChooser.getSelectedFile() +
-                                        ".bin' сохранен");
-                    } else {
-                        throw new IOException();
+                String fileName = saveChooser.getSelectedFile() + ".bin";
+                int flag = 1;
+                File file = new File(fileName);
+                if (file.exists()) {
+                    flag = -1;
+                    int ind = JOptionPane.showConfirmDialog(this, "Файл с таким названием уже существует в данном расположении. Вы хотите сохранить файл с названием " +
+                                    HelperMethods.getFinalNewDestinationFile(new File("output"), file).getName() + "?",
+                            "Предупреждение", JOptionPane.YES_NO_OPTION);
+                    if (ind == 0) {
+                        file = HelperMethods.getFinalNewDestinationFile(new File("output"), file);
+                        flag = 1;
                     }
-                } catch (IOException e) {
-                    new ErrorWindow(this, e);
                 }
+                if (flag != -1){
+                    try (BufferedOutputStream out = new BufferedOutputStream(new FileOutputStream(file))) {
+                        if (function3 != null) {
+                            FunctionsIO.serialize(out, function3);
+                            JOptionPane.showMessageDialog(this,
+                                    "Файл '" + file.getName() +
+                                            " сохранен");
+                        } else {
+                            throw new IOException();
+                        }
+                    } catch (Exception e) {
+                        new ErrorWindow(this, e);
+                    }
+                }
+
             }
+
         });
 
         clearTableButton.addActionListener(evt ->
